Include DigitStack class in index.js
This class will be used for the digit input. Everytime a user presses
a button a digit will be added to the stack from the right, on 'delete'
it will also be removed from the right.

Additionally, for calculations with the call of toNumber, the entire
structure will empty and return itself in number form.