Implement input state machine
Include 4 callsbacks for getInput function to handle each state of
receiving input in the calculator and 1 common assistant-function to
keep things dry.

This is generally not dry, but more like a prototype for testing because
changes might be needed later. Code must remain absolutely
agile during these timestamps.

The code has been tested on some inputs, but not fully. Test-driven
programing at this point is necessary. It will be implemented right
after adding listeners for user input and cleaning the code.

The idea that was implemeneted as stated in previous source code:
/**
My idea is that I will have a singleton which remembers 3 numbers and 3
operators and will only execute operations on 2 numbers at a time.

Scenarios:
1. num.., add/sub, num.., add/sum/equals:
    keep result and operand
2. num.., mult/div, num:
    keep result
3. num.., add/sub, num.., mult/div, num..:
    keep first number, first operand, result of mult/div
 */